Remove crypto features from ssi-dids
And make dependant crates rely on ssi-jwk to enable crypto features
