+++ did-ethr/Cargo.toml
@@ -11,12 +11,8 @@ repository = "https://github.com/spruceid/ssi/"
 homepage = "https://github.com/spruceid/ssi/tree/main/did-ethr/"
 documentation = "https://docs.rs/did-ethr/"
 
-[features]
-ring = ["ssi-dids/ring"]
-openssl = ["ssi-dids/openssl"]
-
 [dependencies]
-ssi-dids = { path = "../ssi-dids", version = "0.1", default-features = false, features = ["secp256k1"] }
+ssi-dids = { path = "../ssi-dids", version = "0.1" }
 ssi-caips = { path = "../ssi-caips", version = "0.1", default-features = false, features = ["eip"] }
 ssi-crypto = { path = "../ssi-crypto", version = "0.1", default-features = false, features = ["secp256k1"] }
 ssi-jwk = { path = "../ssi-jwk", version = "0.1", default-features = false, features = ["secp256k1", "eip"] }
