+++ did-pkh/Cargo.toml
@@ -18,12 +18,9 @@ eip = ["ssi-jwk/eip", "ssi-caips/eip"]
 aleo = ["ssi-jwk/aleo", "ssi-caips/aleo"]
 tezos = ["ssi-jwk/tezos", "ssi-caips/tezos"]
 
-ring = ["ssi-dids/ring"]
-openssl = ["ssi-dids/openssl"]
-
 [dependencies]
 ssi-caips = { path = "../ssi-caips", version = "0.1", default-features = false }
-ssi-dids = { path = "../ssi-dids", version = "0.1", default-features = false }
+ssi-dids = { path = "../ssi-dids", version = "0.1" }
 ssi-jwk = { path = "../ssi-jwk", version = "0.1", default-features = false }
 ssi-crypto = { path = "../ssi-crypto", version = "0.1", default-features = false }
 chrono = { version = "0.4", features = ["serde"] }
