+++ did-sol/Cargo.toml
@@ -11,12 +11,8 @@ repository = "https://github.com/spruceid/ssi/"
 homepage = "https://github.com/spruceid/ssi/tree/main/did-sol/"
 documentation = "https://docs.rs/did-sol/"
 
-[features]
-ring = ["ssi-dids/ring"]
-openssl = ["ssi-dids/openssl"]
-
 [dependencies]
-ssi-dids = { path = "../ssi-dids", version = "0.1", default-features = false }
+ssi-dids = { path = "../ssi-dids", version = "0.1" }
 ssi-jwk = { path = "../ssi-jwk", version = "0.1", default-features = false }
 ssi-caips = { path = "../ssi-caips", version = "0.1", default-features = false }
 chrono = { version = "0.4", features = ["serde"] }
