+++ did-web/Cargo.toml
@@ -11,12 +11,8 @@ repository = "https://github.com/spruceid/ssi/"
 homepage = "https://github.com/spruceid/ssi/tree/main/did-web/"
 documentation = "https://docs.rs/did-web/"
 
-[features]
-ring = ["ssi-dids/ring"]
-openssl = ["ssi-dids/openssl"]
-
 [dependencies]
-ssi-dids = { path = "../ssi-dids", version = "0.1", default-features = false }
+ssi-dids = { path = "../ssi-dids", version = "0.1" }
 async-trait = "0.1"
 reqwest = { version = "0.11", features = ["json"] }
 http = "0.2"
@@ -27,6 +23,7 @@ version = "0.11"
 features = ["json", "native-tls-vendored"]
 
 [dev-dependencies]
+# ssi-jwk = { path = "../ssi-jwk", version = "0.1", default-features = false }
 ssi = { version = "0.4", path = "../", default-features = false }
 tokio = { version = "1.0", features = ["macros"] }
 serde = { version = "1.0", features = ["derive"] }
