+++ did-webkey/Cargo.toml
@@ -13,13 +13,14 @@ documentation = "https://docs.rs/did-webkey/"
 
 [features]
 default = ["secp256r1"]
-secp256r1 = ["ssi-ssh/secp256r1", "ssi-dids/secp256r1"]
+secp256r1 = ["ssi-ssh/secp256r1", "ssi-jwk/secp256r1"]
 
-ring = ["ssi-dids/ring"]
-openssl = ["ssi-dids/openssl"]
+ring = ["ssi-jwk/ring"]
+openssl = ["ssi-jwk/openssl"]
 
 [dependencies]
-ssi-dids = { version = "0.1", path = "../ssi-dids", default-features = false, features = ["http"] }
+ssi-dids = { version = "0.1", path = "../ssi-dids" , features = ["http"] }
+ssi-jwk = { version = "0.1", path = "../ssi-jwk" , default-features = false }
 ssi-ssh = { version = "0.1", path = "../ssi-ssh" }
 anyhow = "1.0.52"
 async-trait = "0.1.52"
