+++ ssi-jwt/Cargo.toml
@@ -29,8 +29,8 @@ ring = ["ssi-jwk/ring", "ssi-jws/ring"]
 thiserror = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-ssi-jwk = { path = "../ssi-jwk", version = "0.1"}
-ssi-jws = { path = "../ssi-jws", version = "0.1"}
+ssi-jwk = { path = "../ssi-jwk", version = "0.1", default-features = false }
+ssi-jws = { path = "../ssi-jws", version = "0.1", default-features = false }
 
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
